pyproject.toml
[project]
name = "easy_agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.0",
    "asyncio>=3.4.3",
    "claude-code-sdk>=0.0.14",
    "fastapi[standard]>=0.116.0",
    "fastmcp>=2.10.4",
    "openai>=1.0.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]

[dependency-groups]
dev = [
    "ruff>=0.12.4",
]
